authorization implemented but there are still bugs

+++ src/api/notes/handler.js
@@ -13,9 +13,12 @@ class NotesHandler {
     async postNoteHandler(request, h){
         this._validator.validateNotePayload(request.payload);
         const { title = 'untitled', body, tags } = request.payload;
-            
-        const noteId = await this._service.addNote({ title, body, tags });
-            
+        const { id: credentialId } = request.auth.credentials;
+
+        const noteId = await this._service.addNote({
+            title, body, tags, owner: credentialId
+        });
+        
         const response = h.response({
             status: 'success',
             message: 'Catatan berhasil ditambahkan',
@@ -27,18 +30,22 @@ class NotesHandler {
         return response;
     }
 
-    async getNotesHandler(){
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request) {
+        const { id: credentialId } = request.auth.credentials;
+        const notes = await this._service.getNotes(credentialId);
         return {
-            status: 'success',
-            data: {
-                notes,
-            },
+          status: 'success',
+          data: {
+            notes,
+          },
         };
-    }
+      }
 
     async getNoteByIdHandler(request, h){
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
+
+        await this._service.verifyNoteOwner(id, credentialId);
         const note = await this._service.getNoteById(id);
         return {
             status: 'success',
@@ -48,11 +55,13 @@ class NotesHandler {
         };
     }
 
-    putNoteByIdHandler(request, h){
+    async putNoteByIdHandler(request, h){
         this._validator.validateNotePayload(request.payload);
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
 
-        this._service.editNoteById(id, request.payload);
+        await this._service.verifyNoteOwner(id, credentialId);
+        await this._service.editNoteById(id, request.payload);
         
         return {
             status: 'success',
@@ -60,9 +69,12 @@ class NotesHandler {
         };
     }
 
-    deleteNoteByIdHandler(request, h) {
+    async deleteNoteByIdHandler(request, h) {
         const { id } = request.params;
-        this._service.deleteNoteById(id);
+        const { id: credentialId } = request.auth.credentials;
+
+        await this._service.verifyNoteOwner(id, credentialId);
+        await this._service.deleteNoteById(id);
         
         return {
           status: 'success',
